refactor: divide code across multiple files
